+++ changelog.md
@@ -5,6 +5,7 @@
 - Added Akelus and M'tara
 - Fixed several sprites that were an odd number of pixels wide from rendering with artifacts due to not being aligned to the pixel grid
 - Fixed Chain Driver's chains remaining player-controlled when switching guns while tethered to an enemy
+- Banned Entropynnium from Blessed Runs
 
 ## 1.31.3 (2026-06-02)